Authorizer core test cases and some refactoring
