+++ schema-registry/authorizer/authorization-agent/src/test/java/com/hortonworks/registries/schemaregistry/authorizer/agent/util/TestAuthorizer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hortonworks.registries.schemaregistry.authorizer.test;
+package com.hortonworks.registries.schemaregistry.authorizer.agent.util;
 
 import com.hortonworks.registries.schemaregistry.authorizer.core.Authorizer;
